python-extractor/requirements.txt
Flask==2.2.5
PyMuPDF==1.22.5
python-docx==0.8.11
pdfminer.six==20221105
